Put the debugging command for banishing into the Angband 4 command system.
