+++ src/cmd-core.h
@@ -103,6 +103,7 @@ typedef enum cmd_code {
 	/* Debugging commands */
 	CMD_WIZ_ACQUIRE,
 	CMD_WIZ_ADVANCE,
+	CMD_WIZ_BANISH,
 	CMD_WIZ_CREATE_ALL_ARTIFACT,
 	CMD_WIZ_CREATE_ALL_ARTIFACT_FROM_TVAL,
 	CMD_WIZ_CREATE_ALL_OBJ,
